rename package to chat.delta.androidyggmail
