+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.cyberta">
+    package="chat.delta.androidyggmail">
 
     <queries>
         <!-- Specific apps you interact with, eg: -->
@@ -62,7 +62,7 @@
 
         <provider
             android:name="androidx.core.content.FileProvider"
-            android:authorities="org.cyberta.fileprovider"
+            android:authorities="chat.delta.androidyggmail.fileprovider"
             android:grantUriPermissions="true"
             android:exported="false">
             <meta-data
