+++ app/src/main/java/chat/delta/androidyggmail/YggmailApp.java
@@ -1,0 +1,14 @@
+package chat.delta.androidyggmail;
+
+import android.app.Application;
+
+public class YggmailApp extends Application {
+
+    private YggmailOberservable yggmailOberservable;
+
+    @Override
+    public void onCreate() {
+        super.onCreate();
+        this.yggmailOberservable = YggmailOberservable.getInstance();
+    }
+}
