--- app/src/main/java/org/cyberta/YggmailApp.java
@@ -1,14 +1,0 @@
-package org.cyberta;
-
-import android.app.Application;
-
-public class YggmailApp extends Application {
-
-    private YggmailOberservable yggmailOberservable;
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-        this.yggmailOberservable = YggmailOberservable.getInstance();
-    }
-}
