+++ app/src/main/res/layout/fragment_settings.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingHorizontal="@dimen/text_margin"
-    tools:context=".settings.SettingsFragment">
+    tools:context="chat.delta.androidyggmail.settings.SettingsFragment">
 
     <androidx.appcompat.widget.SwitchCompat
         android:id="@+id/switch_enable_onboot"
