+++ app/src/main/res/menu/menu_main.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="org.cyberta.MainActivity">
+    tools:context="chat.delta.androidyggmail.chat.delta.AndroidYggmail.MainActivity">
     <item
         android:id="@+id/action_logs"
         android:orderInCategory="100"
